% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spruce.df}
\alias{spruce.df}
\title{Spruce tree height and diameter data}
\format{
A data frame with 36 rows and 2 variables:
\describe{
\item{BHDiameter}{Breast height diameter, in cm.}
\item{Height}{Tree height, in m.}
}
}
\source{
MATH 4753, University of Oklahoma, Lab 3.
}
\usage{
spruce.df
}
\description{
Height and breast height diameter measurements for 36 spruce trees,
as used in MATH 4753 Lab 3 (see MS 10.52).
}
\keyword{datasets}
